requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
cffi==2.0.0
Django==6.0
django-axes==8.3.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==3.0
PyJWT==2.10.1
sqlparse==0.5.4
